[ch210] Handle redirect after the form is completed
