+++ locales/en/sign-up.json
@@ -16,9 +16,9 @@
   "firstNamePlaceholder": "Enter your first name",
   "lastNamePlaceholder": "Enter your last name",
   "inputImproperName": "Please use proper name.",
-  "usernameCharactersMinimum": "Username must contain at least 8 characters.",
+  "usernameCharactersMinimum": "Username must contain at least 6 characters.",
   "usernameAlreadyUsed": "Your username has already been used. Please use a different one.",
   "emailIncorrectFormat": "Your email is incorrect. Please recheck the email format or use a different one.",
   "passwordCharactersMinimum": "Password must contain at least 8 characters.",
-  "passwordRequiredFormat": "Password must contain at least 1 lowercase, 1 uppercase character and 1 number."
+  "passwordRequiredFormat": "Password must contain at least 1 uppercase, 1 special character and 1 number."
 }
